add to circle yml
