--- scripts/deploy.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-echo deploying
