+++ webpack/webpack.config.base.js
@@ -1,0 +1,30 @@
+const CopyPlugin = require('copy-webpack-plugin')
+const path = require('path')
+
+const outputPath = path.resolve(__dirname, '..', 'artifacts', 'webpack_build')
+
+module.exports = {
+    mode: 'development',
+    entry: './src/index.js',
+    output: {
+        filename: 'yvm.js',
+        path: outputPath,
+    },
+    target: 'node',
+    plugins: [new CopyPlugin(['src/yvm.sh'])],
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        cacheDirectory: true,
+                        presets: ['@babel/preset-env'],
+                    },
+                },
+            },
+        ],
+    },
+}
